tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/021.html

<!DOCTYPE html>
<html><head>
  <title> scheduler: DOM added script with javascript: URL </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<body>

  <div id="log">FAILED (This TC requires JavaScript enabled)</div>
  <script>log('inline script #1');
    testlib.addScript('', { 'src':'javascript:log("JS URL script")' }, document.getElementsByTagName('body')[0], true);
    log('end script #1');
  </script>

  <script type="text/javascript">
  log( 'inline script #2' );
  var t = async_test()

  function test() {
    assert_array_equals(eventOrder, ['inline script #1', 'end script #1', /*'JS URL script',*/ 'inline script #2']);
    t.done();
    /* pass condition changed 2010-12-01 due to CT-198 */
  }
  onload = t.step_func(function(){setTimeout( test, 100); })
  </script>

</body></html>
